Adding a C# WS client library for Buttplug
This refactors the shared code into its own project so that it can be shared between the server and client.

The test demonstrates a basic flow

Note: this lacks error handling, unexpected incoming messages and more. But it is a starting point, and much closer than if someone had to write this from scratch.